requirements.txt
# Production requirements
psycopg2==2.9.5
django==4.1.3
gunicorn~=20.1.0
django-environ~=0.9.0
dj3-cloudinary-storage~=0.0.6
django-summernote~=0.8.20.0
django-allauth~=0.51.0
